fix build fiq_glue.h gpio.h do_each_thread fiq_tty_write walk_stackframe

+++ drivers/soc/rockchip/fiq_debugger/fiq_debugger.c
@@ -1208,7 +1208,7 @@ void fiq_tty_wake_up(struct platform_device *pdev)
 }
 EXPORT_SYMBOL_GPL(fiq_tty_wake_up);
 
-static int fiq_tty_write(struct tty_struct *tty, const unsigned char *buf, int count)
+static int fiq_tty_write(struct tty_struct *tty, const unsigned char *buf, size_t count)
 {
 	int i;
 	int line = tty->index;
